+++ src/bindings/python/flux/job/validator/validator.py
@@ -8,16 +8,13 @@
 # SPDX-License-Identifier: LGPL-3.0
 ###############################################################
 
-import os
-import sys
 import json
-import importlib
-import pkgutil
 import argparse
 import threading
 import concurrent.futures
 from abc import ABC, abstractmethod
 import flux
+from flux.importer import import_plugins, import_path
 
 
 class ValidatorResult:
@@ -123,43 +120,6 @@ class ValidatorPlugin(ABC):  # pragma: no cover
             None or (errnum, errmsg) tuple.
         """
         raise NotImplementedError
-
-
-def import_plugins_pkg(ns_pkg):
-    """Import all modules found in the namespace package ``ns_pkg``"""
-    return {
-        name: importlib.import_module(f"{ns_pkg.__name__}.{name}")
-        for finder, name, ispkg in pkgutil.iter_modules(ns_pkg.__path__)
-    }
-
-
-def import_plugins(pkg_name, pluginpath=None):
-    """Load plugins from a namespace package and optional additional paths
-
-    A plugin in pluginpath with the same name as an existing plugin will
-    take precedence
-    """
-    if pluginpath is not None:
-        sys.path[1:1] = pluginpath
-
-    #  Load 'pkg_name' as a namespace plugin
-    pkg = importlib.import_module(pkg_name)
-    plugins = import_plugins_pkg(pkg)
-
-    if pluginpath is not None:
-        #  Undo any added pluginpath elements.
-        for path in pluginpath:
-            sys.path.remove(path)
-
-    return plugins
-
-
-def import_path(file_path):
-    module_name = os.path.basename(file_path).rstrip(".py")
-    spec = importlib.util.spec_from_file_location(module_name, file_path)
-    module = importlib.util.module_from_spec(spec)
-    spec.loader.exec_module(module)
-    return module
 
 
 # pylint: disable=too-many-instance-attributes
